refactor code. Set the hole form as a widget
